Add 215-test suite, fix 3 autograd/dtype bugs in model and diffusion
Test suite (11 files in tests/):
- 52 model tests (forward pass, shapes, gradient flow for all 6 model classes + 7 losses)
- 18 regression tests explicitly named after fixed bugs M1-M5, X1-X3, T1-T2, E1
- 10 training integration tests (train_one_epoch, validate, scheduler, early stopping)
- 15 HyperIV, 19 DDPM, 17 DGM, 21 adjustment, 16 structural break, 14 dataset, 22 utils
- Synthetic data + mock I/O (no real files needed), tiny models for ~4s runtime

Bug fixes found during testing:
- SmileModel.forward: 2nd autograd.grad missing create_graph=True (freed graph prematurely)
- SmileModel.forward: in-place requires_grad=True leaked graph across batches (use detach)
- diffusion.py: t.float() produced float32 in float64 context (use t.to(dtype))

+++ src/diffusion.py
@@ -30,7 +30,7 @@ class SinusoidalTimeEmbedding(nn.Module):
         freqs = torch.exp(
             -math.log(10000) * torch.arange(half, dtype=t.dtype, device=t.device) / half
         )
-        args = t.unsqueeze(-1).float() * freqs.unsqueeze(0)
+        args = t.unsqueeze(-1) * freqs.unsqueeze(0)
         return torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
 
 
@@ -273,7 +273,7 @@ class DiffusionTrainer:
         x_noisy = sqrt_ab * x_target + sqrt_omab * noise
 
         # Predict noise
-        noise_pred = self.model(x_noisy, t.float(), current_surface, market_features)
+        noise_pred = self.model(x_noisy, t.to(x_target.dtype), current_surface, market_features)
 
         # MSE loss on noise prediction
         loss = F.mse_loss(noise_pred, noise)
